Rename previewImportWizard -> importPreview
